fix(clob): send side as string BUY/SELL and salt as integer in order payload
The Polymarket CLOB API expects:
- side: "BUY" or "SELL" (string), not 0/1 (integer)
- salt: JSON integer, not decimal string

Also reduce salt to u32 range to avoid JSON precision loss on large u64 values.

+++ src/clob/signing.rs
@@ -289,7 +289,7 @@ pub async fn sign_order(order: &mut Order, private_key: &str, chain_id: u64) -> 
     order.signer = Some(signer);
 
     if order.salt.is_zero() {
-        order.salt = U256::from(rand::random::<u64>());
+        order.salt = U256::from(rand::random::<u32>());
     }
 
     let typed = order_typed_data(order, maker, signer, chain_id)?;
